Move document symbol tests to separate module
